pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "phishing-analyzer-prod"
version = "1.0.0"
description = "Offline-first phishing email analyzer with optional VirusTotal enrichment"
readme = "README.md"
requires-python = ">=3.11"

authors = [
  { name = "Manish", email = "manishreddy.gutha@gmail.com" }
]

license = { text = "MIT" }

keywords = [
  "phishing",
  "email-security",
  "cybersecurity",
  "streamlit",
  "virustotal"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Security",
  "Topic :: Internet :: Email"
]

dependencies = [
  "streamlit==1.35.0",
  "requests==2.32.3",
  "python-dotenv==1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "pytest-cov==5.0.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov=phishing_analyzer --cov-fail-under=70"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "phishing_analyzer/logging_config.py",
    "phishing_analyzer/agents/header_agent.py"
]
branch = true
source = ["phishing_analyzer"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.setuptools.packages.find]
where = ["."]
include = ["phishing_analyzer*"]
